fix(node): fix stackitem deserializing, headerHashList saving
