+++ packages/neo-one-node/src/__tests__/blockchain.test.ts
@@ -1,12 +1,14 @@
 import { common } from '@neo-one/client-common';
 import { Blockchain } from '@neo-one/node-blockchain';
-import { NativeContainer } from '@neo-one/node-native';
+import { StorageKey, StreamOptions } from '@neo-one/node-core';
+import { KeyBuilder, NativeContainer } from '@neo-one/node-native';
 import { test as createTest } from '@neo-one/node-neo-settings';
 import { storage as levelupStorage } from '@neo-one/node-storage-levelup';
 import { blockchainSettingsToProtocolSettings, Dispatcher } from '@neo-one/node-vm';
 import { BN } from 'bn.js';
 import LevelUp from 'levelup';
 import RocksDB from 'rocksdb';
+import { map, toArray } from 'rxjs/operators';
 import { data } from '../__data__';
 
 describe('blockchain persists second block', () => {
@@ -25,6 +27,8 @@ describe('blockchain persists second block', () => {
       protocolSettings: blockchainSettingsToProtocolSettings(blockchainSettings),
     });
 
+    // console.log(dispatcher.test());
+
     const native = new NativeContainer(blockchainSettings);
     const blockchain = await Blockchain.create({
       settings: blockchainSettings,
@@ -33,6 +37,22 @@ describe('blockchain persists second block', () => {
       native,
     });
 
-    await blockchain.persistBlock({ block: data.debugBlock });
+    await new Promise((resolve, reject) =>
+      db
+        .createReadStream()
+        .on('data', (data: any) => {
+          console.log(`key: ${data.key.toString('hex')}`);
+          console.log(`value: ${data.key.toString('hex')}`);
+        })
+        .on('error', (error) => {
+          reject(error);
+        })
+        .on('close', resolve)
+        .on('end', resolve),
+    );
+
+    // console.log(blockchain.currentBlockIndex);
+
+    // await blockchain.persistBlock({ block: data.debugBlock });
   });
 });
